victara: Fix apps2sd to usb storage
While this may not be a very good idea to use (do you really want
to have a usb stick sticking out of your phone??), the current
fstab is incorrect and needs to be fixed.

Removable storage is always noemulatedsd (and defined as such
in storage_list.xml).  The fstab should be consistent with the
rest of android.

Change-Id: Ic311b7124e325f9ccc8f41889f8ad9b74bc3fb5b