victara: Rearrange partition management
* Move mounts to fstab (aside from pds)
* Remove unused customize partition
* Mark cache and data as formattable
* Handle partition permissions in ueventd

The pds partition has intentionally been left alone here. It will be
handled in a separate commit.

The modem partition is handled slightly different in fstab than in
init. The distinction is that fstab does not attempt to resize the
filesystem on the partition. This should not be any cause for concern
as CyanogenMod OTAs will never alter this partition or filesystem, so
any resizing of the filesystem that is necessary should have already
been done by the stock ROM before coming to CyanogenMod.

Change-Id: I965c416fb4fd9016ab1811d7806255a042876ed7